move shared components to shared folder
